Early termination for prefix doubling
Stop if the order no longer change.  No longer store the whole matrix
of names (because we don't need them anyway, the LCP is computed using
only the suffix array and its inverse).

With this change, the prefix doubling version outperforms the brute force
version at size of around 1000 for random strings.  For strings of the
same character all over it is slower, but that is not an interesting
case.  So the version used by default is changed to the prefix doubling
version.

Signed-off-by: Isaac To <isaacto@gmail.com>